requirements.txt
fastapi
uvicorn
python-dotenv
ollama
openai
pandas
numpy
scipy
polars